[controller] Apply SemanticDetector inside AdminOperationSerializer to detect bad usage for new semantics (#1615)
This PR is a follow-up PR for PR#1574. This PR applies the class SemanticDetector inside 
AdminOperationSerializer to fail the serialization step when new semantic usage is detected.

With this change, when the admin version pinning in ZK is not latest and we receive admin operation
message that using non-default value, we fail the request and return 400 bad request response.
We also introduce a new test integrationTest class to test all operations with previous version to
early detect any bad usage for new semantics. It is expected to add new integration test for new
admin operation message.