[changelog] Add back synchronization around poll()
It seems that our pubsub consumer adapter doesn't even tolerate concurrent calls to poll, so removing locking on this function is pointless.  So we add back an exclusive lock on the poll and in order to reduce the pause time on the version swap seek we shrink the scope of the read lock on version swap to just the short part.