[changelog][client] Refine version swap seeking
When a version swap happens today we seek the cosnumer to the point of EOP and do poll/filtering. This has the advantage of being correct in all cases, but has the downside of meaning a potential increase in delay while the filtering goes on.  This change introduces the notion that the timestamp in the leader produced heartbeat is guaranteed to be earlier then any broker could have received the message on any version.  Using that property to our advantage, we can check and see if for a given EOP message, does this message come before or after the message closest to the timestamp on the last received heartbeat for the current version.  If the EOP message comes after, then it means the consumer is on some checkpoint that is too stale/lagged and EOP message should actually be used.  If the EOP message come before, then it means that everything after that is RT data which was already consumed and can be filtered.  So we'll seek ahead.