[dvc][server][controller][cc][doc][test] DVRT CDC MVP (#1590)
This PR introduces a new implementation of BootstrappingVeniceChangelogConsumer that leverages blob transfer,
enabling a CDC client to bootstrap its state from another peer. The Da Vinci Client (DVC) and the Da Vinci Record
Transformer (DVRT) are utilized in conjunction to achieve this.

Key changes:

1. Version swap messages (VSM) are now being emitted to the Version Topic (VT). Previously, they were only produced
to the Real Time Topic (RT). This was done so DVRT CDC knows when to perform a version swap by consuming the VSM
from the future VT. For hybrid stores, the VSM will be produced to the RT by the child controller, then the leader will
consume it who will produce it to the VT. For batch stores, the child controller will directly produce the VSM to the VT. To
not break the current CDC client, a check was added to ignore VSMs consumed from a VT, as it currently relies on
consuming them from the change-capture topic.

2. Modified ChangelogClientConfig to make using this new client configurable. For the sake of testing, blob transfer is
also configurable. This allows us to compare the performance of when blob transfer is enabled or not. These configs will
eventually be removed in the future once DVRT CDC has stabilized and has become the default.

3. Introduced a new configuration parameter, maxBufferSize, to ChangelogClientConfig. DVRT CDC maintains a message
buffer that users need to drain. To prevent an OutOfMemory error, we set a maximum size on this buffer. The default
setting might not be suitable for all use cases, so it must be configurable.

4. Made the DaVinciRecordTransformer interface partition aware. Additionally, added a new api to
BlockingDaVinciRecordTransformer, onVersionSwap. This will get triggered when a VSM is consumed from the future VT
by DVC. These are necessary to know whether a specific partition is ready to serve.

5. Added a new config skipCompatibilityChecks to DaVinciRecordTransformerConfig. If a user isn't modifying the
contents of the records and is making frequent changes to their DVRT implementation, setting this to true will prevent
DVRT from wiping local state everytime a change is deployed. DVRT CDC is setting this to true due to these reasons.

6. Updated DVRT docs to make the examples partition aware.

7. Modified the setup in TestBootstrappingChangelogConsumer to use a single region as the tests don't depend on it
being multi-region. This was done to simplify the setup logic. Additionally, created helper functions to reduce code
duplication.

8.Turned off compression in testRecordTransformerChunking to ensure chunking threshold is hit.