[vpj] Make Spark job exceptions be thrown during execution instead of configuration (#1513)
In Spark ComputeJob, some cases of "DATASET_CHANGED" were not getting tracked correctly, and were instead getting tracked as "START_DATA_WRITER_JOB".

This was because of an incorrect assumption that Spark jobs would get executed only when the final action stage is encountered. But in practice, the spark stages were starting to get executed when the "repartitionAndSortWithinPartitions" call was getting invoked. This was previously done in the "configure" phase, and exceptions thrown in this stage are not handled as errors during the execution of the compute job.

This commit moves all Spark execution logic inside the "runComputeJob" method instead.

In addition to this, this commit also moves all other input validation inside the same function too so that they can be reported with valid status.