[server] Do not track leader lag in when STANDBY->LEADER is paused due to migration (#1673)
During migration, dest cluster leader is not actually promoted to leader    
and is still replicating local VT. During this stage, we should not track    
leader lag as it is not valid. We should track follower lag, and once this    
stage is lifted, we can track leader lag.    

There are two ways to fix it:    

1. Remove leader lag tracking when transition is paused and resume it after    
2. Just move the whole leader tracking when it is fully transitioned into    
   leader    

In this PR we change to (2), as follower will already be tracking leader    
replication lag, so when leader is truly stuck during transition, follower    
lag will be going up and alarm.    

Also, this PR fixes a small bug when calculating the max follower lag. It    
was using the leader map and changed to follower map. It is not used today    
so no current impact.