[server][misc] Move leader->follower heartbeat tracking action to after transition complete (#1681)
The bug happens is:    
When a leader transits into follower, we will first clear leader HB entry    
and initialize follower HB entry, and then submit request to transition into    
follower to KafkaStoreIngestionService.    

The problem is, in between, the leader will need to drain all the consumed    
messages before turning into follower and start consuming VT. In between, if    
we see any leader HB message, it will be recorded into the leader entry, but    
then this will remain as a stale leader heartbeat and get recorded.