[server][common] close MaterializedViewWriter properly in LFSIT (#1684)
* [server][common] close MaterializedViewWriter properly in LFSIT

MaterializedViewWriter did not implement the close() method defined in VeniceView. This means
when SIT's internalClose is called the VeniceWriter in MaterializedViewWriter will not be
closed. This can cause stuck consumer because a consumer thread could be waiting for the
last VT produce future in checkAndWaitForLastVTProduceFuture to ensure CMs that don't need
to be produced to view(s) will not be produced to VT out of order. Without closing/flushing
the view writer and potentially shortcircuit the lastVTProduceCallFuture the consumer thread
could be waiting on a lastVTProduceCallFuture forever and here is one example how this could
occur with the old code:
  1. For each incoming record we get the lastVTProduceCallFuture to be part of a new future
     that is completed when the view writers for the current record and lastVTProduceCallFuture
     are completed. This new future is then set as the new lastVTProduceCallFuture.
  2. This chaining can go for many layers. i.e.
     {[(lastVTPCF1, viewWriterFuture1), viewWriterFuture2], viewWriterFuture3}
     The consumer thread could be waiting for the lastVTProduceCallFuture defined with { }.
     However, VT VW/producer could be closed already when we try to produce to VT in the callback
     of the lastVTProduceCallFuture defined with ( ). This will cause the handling thread,
     currently the common ForkJoinPool to terminate because version topic write is a runnable.
  3. This means the lastVTProduceCallFuture defined in [ ] will never complete and the same goes
     with the future defined with { } which depends on [ ] and ( ). Consumer thread waiting on
     { } will be stuck forever.

The change closes the view writers first, wait or shortcircuit any incomplete lastVTProduceCallFuture
and then close the VT writers.