[server][dvc] Improve disk healthcheck error reporting (#1691)
In case when disk health check fails due to seeing a stale status update, the logs only contain:
"Disk is not healthy according to the disk health check service: null"
This does not provide enough information about the issue

This commit adds a reasonable error message in this case.

To avoid running into this issue again, this commit also changes the error tracking to always require an error message when marking the disk health as unhealthy. It also ensures to clean up the error message when the disk is marked healthy.