[controller] Fast leader to follower state transition for controller
We have observed that for Venice controller, the slow leader-to-standby transition is problematic.
This is because this controller is still in the leadership role and accepting requests as
the transition is in progress. However, during this transition, its internal metadata is in the
process of being cleared and when it replies to any requests, it could result in 404 or store
not found errors, while the store does exist.

The solution is to implement a fast leader-to-standby state transition for the controller. We can
delay the work of clearing resources by using a single worker thread and let the state transition
thread complete quickly. For standby to leader state transition, however, we can still handle it
slowly, that is only declare the leadership when all preparation work is done.

Note that, we need to pay attention that if a future standby to leader state transition happens,
it has to be put into the same worker queue so that any previous dependent state transitions
are completed first.