[controller] Report detailed reason for failing an instance's stoppability request (#1714)
Today, in Helix's aggregated stoppable check call ("/aggregatedHealthStatus"), the response when removal is blocked only contains enums like "WILL_LOSE_DATA" or "WILL_TRIGGER_LOAD_REBALANCE". This does not include any details about the actual resource that is blocking the removal making the engineer go through additional hoops to get that information. This change adds the a detailed reason for preventing the removal of a node to the response message