[router] Use Helix FAULT_ZONE_TYPE as the group for HAR routing (#1710)
Today, routers only support "zone" and "group" as the "FAULT_ZONE_TYPE" when
the routing strategy is "HELIX_ASSISTED" (HAR). Assuming "FAULT_ZONE_TYPE" is
set to "zone", routers will look for a property with key "zone" in the Helix
"DOMAIN" of all server instances to find a replica to route requests to.

When the "FAULT_ZONE_TYPE" changes (say to "group"), Helix will immediately
start rebalancing replicas and the constraints we create for HAR are no longer
valid. Currently, the value of "FAULT_ZONE_TYPE" that routers look for is
configured through it's configs. To make routers use the updated
"FAULT_ZONE_TYPE" ("group"), we need to update the configs of the routers and
deploy them. This process is slow and there will be performance degradation
when the value set in router configs does not align with the value being used
by Helix.

This commit changes router logic to use Helix's "FAULT_ZONE_TYPE" as the key in
the "DOMAIN".