[server] Make ST wait for ingestion completion for completed future versions (#1715)
During 3-4-3 helix rebalancing, future versions that are already completed and waiting for a deferred swap some time in the future will undergo fast ST i,e without waiting for the ingestion completion. This will result in the new replicas without servable data (3 bootstraping instances) for sometime until the ingestion is completed and if the deferred swap was called around this time, the replicas will fail serving data until it eventually becomes ready. Fix is to Make server ST for completed future version behave the same way as the current version: wait for ingestion completion by taking a latch to avoid fast ST.