[controller] ProtocolVersionAutoDetectionService to periodically detect the good version among controllers (#1696)
This PR is the follow-up PR for #1669.
The mission is to automate the process to remove deployment order for controllers for Admin Ops
Protocol Upgrade.
1. Allow ControllerClient to send request to a specific host given URL instead of sending to LEADER
2. Setup ProtocolVersionAutoDetectionService that runs every 10 minutes (configurable) to get all
latest admin operation versions from all consumers controllers given cluster, and then find smallest
version to pin in ZK.
This service shares the same lifecycle with HelixVeniceClusterResources and only run within
parent leader controller.