[doc] Enable dark mode for documentation (#1677)
The current theme for the documentation is light theme and it hurts my eyes.

The solution is enabling the dark theme for Just the Docs, which can be viewed here (https://kvargha.com/venice/).

There were also diagrams in vip3 and vip4 that had transparent backgrounds, so I added a white background to make
sure it works with the dark theme.

I've looked into adding a toggle between themes, but it's not very clean.