[dvc][server] Refactored Transient Record usage code (#1744)
There was an inconsistency where the transient record cache (which is
used for WC and AA ingestion) had different criteria for whether to
write and read to it. We would not write to it when pre-EOP (which is
the right behavior), but we would systematically read from it in the
setChunkingInfo function of the LeaderProducerCallback, regardless of
the EOP status.

This commit adds an extra PartitionConsumptionState parameter to the
SIT::isTransientRecordBufferUsed function, and calls this function
in more locations, thus eliminating duplicate conditions littered
across the code.

Miscellaneous:

- Fixed the logger class name in the LeaderProducerCallback.

- Eliminated a few unused partitionId parameters in SIT functions.