[doc] Tweaked main logo so that it looks better on various backgrounds (#1752)
Rather than slapping a big white rectangle with rounded corners behind
the main logo, on the main README page, this commit now instead places
an identical image with thicker white outlines behind the normal image.
This method is chosen rather than just adding outlines to the normal
image, because doing so results in weird artifacts around the borders.

This should look better on dark backgrounds, and still look fine over
white as well (in which case the new border will be invisible).