[vpj] Explicitly detect Vson file and not suppress IOException (#1745)
There is no official utility to detect if a file is a "SequenceFile". We rely on "SequenceFile.Reader" to throw an "IOException" and then we catch this "IOException" and track all "IOException" errors as not being a "SequenceFile". However, this approach masks true "IOException" and eventually, VPJ ends up with a "NullPointerException" through some sequence of events.