[server] Improve ReadQuotaEnforcementHandler init() behavior and visibility
Previously the init() call in ReadQuotaEnforcementHandler is made async due to CV repository
may not be ready by the time when the constructor of ReadQuotaEnforcementHandler is called.
The async init() implementation had several flaws which are addressed in this PR:
  1. Exceptions thrown in init() are not caught and logged resulting in silent failures.
  2. If init() fails or takes too long to finish all requests are failing open without
     invoking recordAllowedUnintentionally which also results in silent failure.
  3. The async init was done by the common fork join pool which comes with some undesirable
     traits like it could be exhausted, no control over external dependencies and etc.

Changes:

1. A single thread executor with DaemonThreadFactory will be used to perform the async init.
The executor is also shutdown after submitting the initialization task so no resource
hogging once we are done with async initializaiton.

2. Requests that are allowed while waiting for ReadQuotaEnforcementHandler initialization are
considered for fail open quota requests and will invoke recordAllowedUnintentionally. Only
requests for stores that do not have storage node quota enabled will be allowed without leaving
any metric trace from the ReadQuotaEnforcementHandler.