[dvc][server] Improve Hybrid Store Validation for Real-Time Topics in TopicExistenceChecker (#1751)
The current implementation of shouldTopicExist in MetadataRepoBasedStaleTopicCheckerImpl
checks if a store is hybrid when validating a real-time topic. However, this check is insufficient
because a version can be hybrid while the store itself is non-hybrid.

This PR is to update the code to validate hybrid version for RT topic instead of using store.isHybrid()