[server] Report ingestion error to controller for batch stores (#1704)
During ingestion, if after EOP server or DVC instance hits any error, it will not be reported.  as an error to the controller. The controller then waits for 24hrs timeout limit to mark the push   as failed blocking further pushes.

This PR reports error for such failed pushes even if EOP is received for batch stores.