[fast-client][router] Disable auto retry in httpclient5 to avoid delay upon 429/503 response
HttpClient5 has a feature: auto retry and check this class to find more details:
{@link org.apache.hc.client5.http.impl.DefaultHttpRequestRetryStrategy}
This feature is enabled by default, when this feature is enabled,
upon 429/503 responses, it will introduce a delay of 1s (default), and retry,
which means the user of this lib: Fast-Client and Router will observe
>1s latency, which is unexpected.
This PR disables this feature to avoid the above behavior.
Whether retry upon these error response or not will be decided by the
user instead of the transporting layer.