[server][dvc] Add duplicate key count metric for rocksDB (#1766)
The version topic collects duplicate keys for writing or delete to the same keys.  
During Kafka log compaction those are they are de-duped and VT size is maintained. 
under control. But without log-compaction, we need to find stores with lots of duplicate. 
writes which can be repushed to trim the VT topic size. This PR adds a metric to find   
the duplicate key count from rocksdb statistics.