[controller] Update the calculation for startProcessing metric (#1781)
adminMessageStartProcessingLatencySensor is the metric to calculate the time that the message
is queuing for process in internalTopic. However, it is recorded as the time from message arrived
to the broker (getLocalBrokerTimestamp) --> the first time it is processing.

This PR updates the calculation based on the delegateTimestamp.

So now we have:

delegate_latency: available in local topic -> delegated
start_processing_latency: delegated -> first processing attempt
process_latency: first processing attempt -> processed