[venice-thin-client] Removed async warmup in Venice Thin Client
The async warmup in venice-thin-client has introduced a lot of
confusions and issues in the past and the hope was to mitigate
the cold start issue when the users don't start the client in
the internal startup aware callback, which is typically this case,
and that was why we introduced async warmup since the dependent
d2 client might not be ready when `start` is invoked and we hope
the startup procedure can be done before sending out the first
request, but in practice, it doesn't hold true for most of the cases.

We discovered a new way to invoke the `start` method of the store
clients in the startup aware callback, and it will be more deterministic
than the best-effort async warmup.
So this PR removes the async warmup completely.

This PR exposes a new method: InternalAvroStoreClient#startWithExceptionThrownWhenFail,
and we will use this one in the proprietary layer.