[server] Add logging for subscribed but inactive topic partitions in consumer service (#1737)
The intention is to detect topic partitions that are subscribed but received no messages for
an extended period of time due to any of the following reasons:
   1. Starvation due to shared consumer
   2. Pub-sub broker or client issues
   3. Code bug
All subscribed topic partitions are expected to receive messages from polls because:
   1. RT topics have heartbeat messages produced to them periodically. VT for hybrid stores will receive them too.
   2. VT for batch only stores should unsubscribe after completion (EOP received).
Currently the tracker is unable to differentiate RT topic subscriptions across different versions. e.g. current
version and future version leader for a given partition is on the same host and subscribed to different offset of the
RT topic. Current version is able to poll successfully and future version cannot. The current implementation is
able to detect but unable to identify which subscription is problematic since there is only one timestamp per topic partition.