[server][dvc] Add support to use heartbeat for ready-to-serve check (#1774)
This PR adds a new mode to use heartbeat to measure ready-to-serve for any    
given replica on the node.    

When the config is enabled, the server/DVC instance will use heartbeat lag    
instead of offset lag to measure if a partition is caught up and    
ready-to-serve.    

This PR also deprecates hybrid store time lag usage, as it is not used    
anywhere and heartbeat lag makes more sense as a replacement.    

Also, there has been some offline discussion about refactoring all the    
offset management. This won't happen in the current PR as the change would    
be too large. It will gradually take place in future related PRs.