[controller] Add metric for stalled version swap (#1775)
Add a new metric to track stalled version swaps. A version swap for a store is considered stalled if the push completion time and more than 110% of the store wait time has passed. If this happens, we will emit a count and decrement this count when the version swap happens for this store.