[vpj] Report empty input source as user error (#1810)
Currently, as part of DefaultInputDataInfoProvider, we throw RuntimeException when the source is empty or invalid. However, we want to treat this as user error instead of an runtime exception.

Solution
Treating invalid input as user error by throwing the appropriate exception, helps VenicePushJob classify the error correctly with the right checkpoint.

Note: DefaultInputDataInfoProvider is an API that can be extended for various input data info providers. We don't have the right structure / abstractions for handle errors as part of this API and as a result, extension of this API may still need to handle invalid input as recommended by this PR. However, we can't enforce consistent handling across the implementations and to enforce that would be out of the scope for this bug fix.