[router] Create a router QPS sensor of type count (#1824)
Problem Statement
Any internal system in LinkedIn is only able to measure QPS when the metric is of type Count. But the only metric we offer today in our router to measure is an OccurenceRate.

Solution
To solve this issue, I created a separate metric to measure router QPS using the sensor type Count.