[router] Rename call_count metric for qps to request_call_count (#1830)
Problem Statement
This PR (https://github.com/linkedin/venice/pull/1824) added a new metric called call_count to measure QPS for the router of type Count. However, in order for this be registered internally as a counter the metric name needs to be formatted in a certain way.

Solution
To fix this, we rename the metric name from call_count to request_call_count.