[server] Delete ignored upstream RT offset tracking from PCS (#1827)
Remove tracking of ignored upstream RT offsets from PartitionConsumptionState (PCS).  
This was previously used to compute lag when a leader ignored messages and there was  
no guarantee of future data. Now that heartbeat messages ensure continuous traffic,  
this mechanism is no longer needed and can be safely removed to simplify the code.