[controller] Fix system store auto-repair service (#1828)
Today, system store auto repair service has the following problems:    

1. It checks the current version of the parent region (which is always 0)    
   and marks all of them as unhealthy.    

2. It uses the wrong controller API to get the largest version ever used in    
   all regions. The API actually returns the largest graveyard version    
   instead.    

This PR addresses the above issues and bumps the default heartbeat wait    
time to 10 minutes. This is to ensure that in the case of a surge in writes,    
the system can still consider itself healthy.    

The wait time is configurable and can be tuned later.