[server] Remove unnecessary sensors from servers topic manager stats (#1829)
This change reduces the number of sensors registered by `TopicManagerStats` for  
the SERVER component by introducing selective registration and removing percentiles.  
Only a predefined set of shared metrics is now tracked on the server side, cutting down  
on unneeded time series reported to monitoring systems.  

Metrics selection for SERVER is based on two factors:  
1. Metrics we've commonly used in the past for debugging server issues.  
2. The server doesn't create topics, so it doesn’t make sense to include sensors  
   related to topic creation—similar reasoning was applied to other excluded sensors.  

As part of this, `VeniceComponent` was converted to an enum to cleanly distinguish  
between CONTROLLER and SERVER usage during stats registration.