[controller] Include store graveyard when computing largest used version for system stores (#1845)
When materializing a system store after deletion and recreation, the user store config  
may not yet contain valid system store entries—since not all configs from the store  
graveyard are applied to the new store. To ensure correctness, we also check the store  
graveyard for the largest used version number when auto materializing the user system  
store.  

This resolves flakiness in  
`testParentControllerAutoMaterializeDaVinciPushStatusSystemStore`, which expects  
system stores to be auto-materialized using the largest version observed across regions.