[dvc][server] Do not fail ingestion when handling regular DELETE when deferred write is enabled (#1839)
Da Vinci ingestion can fail during A/A repush when the sorted key is    
associated with the DELETE operation. Since Da Vinci does not use the    
RMD-based RocksDB implementation, it follows the regular RocksDB DELETE    
path. If deferred write is enabled, this can throw an exception.    

This PR fixes the issue by checking the deferred write status and skipping    
the DELETE when deferred write is true.