[changelog] create metadata repository in change log consumer using helper
Change log consumer is creating a specific type of metadata repository, ThinClientMetaStoreBasedRepository. Eventually we will migrate this to use the new RequestBasedMetaRepository. By using the NativeMetadataRepository.getInstance() helper to create the repository, we will be able to create the correct type using a config. Currently this config is clientConfig.useRequestBasedMetaRepository. We will add a global config in the near future.