[cc] Change CDC count metrics from type Avg to Count (#1853)
Problem Statement
This PR (https://github.com/linkedin/venice/pull/1820) added more metrics to CDC, but the count metrics were registered with the wrong sensor type.

Solution
To fix this, we will register count metrics using Count instead of Avg. The concequences of not doing will result in the count metrics showing 1 all of the time, which is misleading.