[fast-client] Fixed a race condition to avoid wrong replicas (#1866)
Previously, the replcias returned by {@link RequestBasedMetadata#getReplicas}
can be modified by the shuffle operation in {@link LeastLoadedClientRoutingStrategy#getReplicas},
while some other requests are reading the partition assignment.
The List is not thread-safe, so the temp state of shuffle can be read,
which can result in identical replicas assigned to the same partition,
which would break the retry request.

The fix is to return a read-only copy from {@link RequestBasedMetadata#getReplicas}, and
if there is a need to shuffle the read-only list, the user will make a copy to avoid
the race condition.