[vpj] add total uncompressed records exceeding size limit in job details (#1867)
VPJ needs to report record information before compression. We can achieve this using the data collected by the accumulators. The data will be reported to the PushJobDetails system store where they can be exposed.
Information to be exposed includes uncompressed records that are larger than the set limit and the largest uncompressed value size that is recorded in the push job.