[admin-tool] Add support to update largest used RT version in admin tool (#1894)
We ran into cases where store deletion failed as the graveyard had a larger largest used  
RT version than the store being deleted (due to RT versioning rollout, rollback, and  
recreation). We had to manually bump the store's largest used RT version to exceed the  
graveyard's. The admin tool lacked a command for this, so we had to patch it. This CLI  
may help during future rollouts or cleanup scenarios.