[server][dvc] IngestionTaskReusableObjects.Strategy config
The AASIT hangs on to a private property of thread-local type, where each
instance can contain a 1 MB blob. When there are lots of AASIT instances,
accessed from lots of ingestion threads, this can add up to a lot of big
thread-local state.

In a production heap dump with 191 AASIT instances, 51% of all bytes were
occupied by byte[] (though not all of them came from this particular
AASIT field).

With this commit, the default behavior is the same as before, but a new
config allows passing an enum to select alternative behaviors:

server.ingestion.task.reusable.objects.strategy

And the valid values are:

- NO_REUSE: will pass null into the relevant code paths, which will
  result in new objects getting allocated on the fly, and needing to be
  garbage collected.

- THREAD_LOCAL_PER_INGESTION_TASK: This avoids hot path allocations but
  is likely not very efficient , since there can be many AASIT and each
  will have its own independent set of objects, even if/when the AASIT
  becomes dormant. This is the first mode which was built, and it is
  considered stable.

- SINGLETON_THREAD_LOCAL: Likely the most efficient.

Test changes:

- Integration tests override this setting to SINGLETON_THREAD_LOCAL.