[controller][server] Deprecated the Helix Message Channel
The Helix Message Channel is a legacy mechanism for letting the controller
send kill job commands to the servers. It has been replaced a long time ago
by the Participant Store but the config defaults did not reflect that.

This commit aligns the default config with production best practices. A
later commit will delete this code entirely.

Miscellaneous:

- Refactored the flow of the ParticipantStoreConsumptionTask run loop a
  little bit. Added extensive unit testing for it.

- Refactored the VeniceHelixAdmin::close function so that it cannot hang
  forever waiting for the HelixManager to disconnect...

- Added a SleepStallingMockTime to help with multi-threaded testing.