[server][dvc] Check for previous dvc pushstatus retry count (#1892)
topicToNoDaVinciStatusRetryCountMap is a global variable that we use to track how many times we did not receive any  
DVC status report. If we exceed daVinciPushStatusNoReportRetryMaxAttempts, then we remove it from the. 
topicToNoDaVinciStatusRetryCountMap map, and the next call will again start the counter from 1 in a loop.  
This PR check for previous dvc pushstatus query retry count before checking again.